Sujet I/local_webapp_ui/requirements.txt
streamlit>=1.35
# Added for OneNote Graph export (onenote_exporter)
msal>=1.27.0
requests>=2.31.0
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0
